Fix the third-party functionality to work with the new upstream code
