+++ repos/system_upgrade/common/actors/peseventsscanner/actor.py
@@ -51,9 +51,10 @@ class PesEventsScanner(Actor):
         for vendor_list in self.consume(ActiveVendorList):
             active_vendors.extend(vendor_list.data)
 
+        pes_json_suffix = "_pes.json"
         if os.path.isdir(VENDORS_DIR):
-            vendor_pesfiles = list(filter(lambda vfile: ".json" in vfile, os.listdir(VENDORS_DIR)))
+            vendor_pesfiles = list(filter(lambda vfile: pes_json_suffix in vfile, os.listdir(VENDORS_DIR)))
 
             for pesfile in vendor_pesfiles:
-                if pesfile[:-5] in active_vendors:
+                if pesfile[:-len(pes_json_suffix)] in active_vendors:
                     pes_events_scanner(VENDORS_DIR, pesfile)
