+++ repos/system_upgrade/common/libraries/repomaputils.py
@@ -1,154 +1,147 @@
-﻿import os
-import io  # Python2/Python3 compatible IO (open etc.)
+﻿import json
+from collections import defaultdict
 
 from leapp.exceptions import StopActorExecutionError
-from leapp.libraries.common import config
 from leapp.libraries.common.fetch import read_or_fetch
-from leapp.libraries.stdlib import api
-from leapp.models import RepositoriesMap, RepositoryMap, VendorRepositoriesMapCollection
-from leapp.models.fields import ModelViolationError
-
-REPOMAP_FILE = "repomap.csv"
-"""Path to the repository mapping file."""
-BASE_REPOMAP_DIR = "/etc/leapp/files"
-VENDOR_REPOMAP_DIR = "/etc/leapp/files/vendors.d"
+from leapp.models import PESIDRepositoryEntry, RepoMapEntry
 
 
-def _raise_error(msg, details):
+def inhibit_upgrade(msg):
     raise StopActorExecutionError(
         msg,
-        details={
-            "details": details,
-            "hint": (
-                "Read documentation at the following link for more"
-                " information about how to retrieve the valid file:"
-                " https://access.redhat.com/articles/3664871"
-            ),
-        },
-    )
+        details={'hint': ('Read documentation at the following link for more'
+                          ' information about how to retrieve the valid file:'
+                          ' https://access.redhat.com/articles/3664871')})
 
 
-def read_local(
-    filename,
-    directory=BASE_REPOMAP_DIR,
-    allow_empty=False,
-    encoding="utf-8",
-):
-    logger = api.current_logger()
-    local_path = os.path.join(directory, filename)
+def read_repofile(repofile, directory="/etc/leapp/files"):
+    # NOTE: what about catch StopActorExecution error when the file cannot be
+    # obtained -> then check whether old_repomap file exists and in such a case
+    # inform user they have to provde the new repomap.json file (we have the
+    # warning now only which could be potentially overlooked)
     try:
-        with io.open(local_path, encoding=encoding) as f:
-            data = f.read()
-            if not allow_empty and not data:
-                _raise_error(
-                    local_path, "File {} exists but is empty".format(local_path)
-                )
-            logger.warning(
-                "File {lp} successfully read ({l} bytes)".format(
-                    lp=local_path, l=len(data)
-                )
+        return json.loads(read_or_fetch(repofile, directory))
+    except ValueError:
+        # The data does not contain a valid json
+        inhibit_upgrade('The repository mapping file is invalid: file does not contain a valid JSON object.')
+    return None  # Avoids inconsistent-return-statements warning
+
+
+class RepoMapData(object):
+    VERSION_FORMAT = '1.0.0'
+
+    def __init__(self):
+        self.repositories = []
+        self.mapping = {}
+
+    def add_repository(self, data, pesid):
+        """
+        Add new PESIDRepositoryEntry with given pesid from the provided dictionary.
+
+        :param data: A dict containing the data of the added repository. The dictionary structure corresponds
+                     to the repositories entries in the repository mapping JSON schema.
+        :type data: Dict[str, str]
+        :param pesid: PES id of the repository family that the newly added repository belongs to.
+        :type pesid: str
+        """
+        self.repositories.append(PESIDRepositoryEntry(
+            repoid=data['repoid'],
+            channel=data['channel'],
+            rhui=data.get('rhui', ''),
+            repo_type=data['repo_type'],
+            arch=data['arch'],
+            major_version=data['major_version'],
+            pesid=pesid
+        ))
+
+    def get_repositories(self, valid_major_versions):
+        """
+        Return the list of PESIDRepositoryEntry object matching the specified major versions.
+        """
+        return [repo for repo in self.repositories if repo.major_version in valid_major_versions]
+
+    def get_version_repoids(self, major_version):
+        """
+        Return the list of repository ID strings for repositories matching the specified major version.
+        """
+        return [repo.repoid for repo in self.repositories if repo.major_version == major_version]
+
+    def add_mapping(self, source_major_version, target_major_version, source_pesid, target_pesid):
+        """
+        Add a new mapping entry that is mapping the source pesid to the destination pesid(s),
+        relevant in an IPU from the supplied source major version to the supplied target
+        major version.
+
+        :param str source_major_version: Specifies the major version of the source system
+                                         for which the added mapping applies.
+        :param str target_major_version: Specifies the major version of the target system
+                                         for which the added mapping applies.
+        :param str source_pesid: PESID of the source repository.
+        :param Union[str|List[str]] target_pesid: A single target PESID or a list of target
+                                                  PESIDs of the added mapping.
+        """
+        # NOTE: it could be more simple, but I prefer to be sure the input data
+        # contains just one map per source PESID.
+        key = '{}:{}'.format(source_major_version, target_major_version)
+        rmap = self.mapping.get(key, defaultdict(set))
+        self.mapping[key] = rmap
+        if isinstance(target_pesid, list):
+            rmap[source_pesid].update(target_pesid)
+        else:
+            rmap[source_pesid].add(target_pesid)
+
+    def get_mappings(self, src_major_version, dst_major_version):
+        """
+        Return the list of RepoMapEntry objects for the specified upgrade path.
+
+        IOW, the whole mapping for specified IPU.
+        """
+        key = '{}:{}'.format(src_major_version, dst_major_version)
+        rmap = self.mapping.get(key, None)
+        if not rmap:
+            return None
+        map_list = []
+        for src_pesid in sorted(rmap.keys()):
+            map_list.append(RepoMapEntry(source=src_pesid, target=sorted(rmap[src_pesid])))
+        return map_list
+
+    @staticmethod
+    def load_from_dict(data):
+        if data['version_format'] != RepoMapData.VERSION_FORMAT:
+            raise ValueError(
+                'The obtained repomap data has unsupported version of format.'
+                ' Get {} required {}'
+                .format(data['version_format'], RepoMapData.VERSION_FORMAT)
             )
-            return [line.strip() for line in data.splitlines()]
-    except EnvironmentError:
-        _raise_error(
-            local_path, "File {} exists but couldn't be read".format(local_path)
-        )
-    except Exception as e:
-        raise e
 
+        repomap = RepoMapData()
 
-def read_or_fetch_repofile(repofile, directory):
-    contents = read_or_fetch(repofile, directory)
-    return [line.strip() for line in contents.splitlines()]
+        # Load reposiories
+        existing_pesids = set()
+        for repo_family in data['repositories']:
+            existing_pesids.add(repo_family['pesid'])
+            for repo in repo_family['entries']:
+                repomap.add_repository(repo, repo_family['pesid'])
 
+        # Load mappings
+        for mapping in data['mapping']:
+            for entry in mapping['entries']:
+                if not isinstance(entry['target'], list):
+                    raise ValueError(
+                        'The target field of a mapping entry is not a list: {}'
+                        .format(entry)
+                    )
 
-def scan_repomaps(repomap_file, repomap_dir, read_repofile_func=read_or_fetch_repofile):
-    """
-    Scan the repository mapping file and produce RepositoriesMap msg.
-
-    See the description of the actor for more details.
-    """
-    _exp_src_prod_type = config.get_product_type("source")
-    _exp_dst_prod_type = config.get_product_type("target")
-
-    repositories = []
-    line_num = 0
-    for line in read_repofile_func(repomap_file, repomap_dir)[1:]:
-        line_num += 1
-
-        api.current_logger().debug("Grabbing line {} of file {}: \"{}\"".format(line_num, repomap_file, line))
-
-        # skip empty lines and comments
-        if not line or line.startswith("#"):
-            api.current_logger().debug("Line skipped")
-            continue
-
-        try:
-            (
-                from_repoid,
-                to_repoid,
-                to_pes_repo,
-                from_minor_version,
-                to_minor_version,
-                arch,
-                repo_type,
-                src_prod_type,
-                dst_prod_type,
-            ) = line.split(",")
-
-            # filter out records irrelevant for this run
-            if (
-                arch != api.current_actor().configuration.architecture
-                or _exp_src_prod_type != src_prod_type
-                or _exp_dst_prod_type != dst_prod_type
-            ):
-                api.current_logger().debug("Line filtered out")
-                continue
-
-            new_repo_map = RepositoryMap(
-                    from_repoid=from_repoid,
-                    to_repoid=to_repoid,
-                    to_pes_repo=to_pes_repo,
-                    from_minor_version=from_minor_version,
-                    to_minor_version=to_minor_version,
-                    arch=arch,
-                    repo_type=repo_type,
+                for pesid in [entry['source']] + entry['target']:
+                    if pesid not in existing_pesids:
+                        raise ValueError(
+                            'The {} pesid is not related to any repository.'
+                            .format(pesid)
+                        )
+                repomap.add_mapping(
+                    source_major_version=mapping['source_major_version'],
+                    target_major_version=mapping['target_major_version'],
+                    source_pesid=entry['source'],
+                    target_pesid=entry['target'],
                 )
-
-            api.current_logger().debug("Map added: {}".format(new_repo_map.dump()))
-            repositories.append(new_repo_map)
-
-        except (ModelViolationError, ValueError) as err:
-            _raise_error(
-                "The repository mapping file is invalid. It is possible the file is out of date.",
-                "Offending line number: {} ({}).".format(line_num, err),
-            )
-
-    if not repositories:
-        _raise_error(
-            "The repository mapping file is invalid. Could not find any repository mapping record.",
-            "",
-        )
-
-    return RepositoriesMap(file=repomap_file, repositories=repositories)
-
-
-def scan_vendor_repomaps(repomap_dir):
-    if not os.path.isdir(repomap_dir):
-        api.current_logger().debug(
-            "The {} directory doesn't exist. Nothing to do.".format(repomap_dir)
-        )
-        return None
-
-    vendor_maps = []
-
-    for repomap_name in os.listdir(repomap_dir):
-        # Only scan the .csv files, those are the maps.
-        if not repomap_name.endswith(".csv"):
-            continue
-        scanned_map = scan_repomaps(
-            repomap_name, repomap_dir, read_repofile_func=read_local
-        )
-        vendor_maps.append(scanned_map)
-
-    return VendorRepositoriesMapCollection(maps=vendor_maps)
+        return repomap
